1008번 문제 5차 시도 | https://www.acmicpc.net/problem/1008

+++ problems/1008.c
@@ -15,7 +15,7 @@ int main() {
     int a;
     int b;
     scanf("%d %d", &a, &b);
-    float c = a / b;
-    printf("%d", c);
+    float c = (double) a / b;
+    printf("%lf", c);
     return(0);
 }
